% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTML.R
\name{sfp_fullpage}
\alias{sfp_fullpage}
\alias{sfp_section}
\alias{sfp_slide}
\title{HTML Wrappers}
\usage{
sfp_fullpage(id, ..., options = NULL)

sfp_section(...)

sfp_slide(...)
}
\arguments{
\item{id}{HTML id}

\item{...}{Additional parameters passed onto \code{div}}

\item{options}{List of options see notes}
}
\description{
HTML Wrappers
}
\note{
\url{}
}
