% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include.R
\name{sfp_include}
\alias{sfp_include}
\alias{sfp_lib}
\title{Include fullPage.js library}
\usage{
sfp_include(cdn = TRUE)

sfp_lib(cdn = TRUE)
}
\arguments{
\item{cdn}{Boolean indicate whether to pull files from CDN}
}
\description{
Include fullPage.js library
}
